refactor: Update font size and color in PatientDetailStyles and PatientsAllStyles
This commit updates the font size and color in the `PatientDetailStyles` and `PatientsAllStyles` files. In `PatientDetailStyles`, the font size of the `cellText` class is increased to 17, and the color is set to '#333'. In `PatientsAllStyles`, the font size of the `cellText` class is also increased to 17. These changes improve the consistency and readability of the codebase.